+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         </activity>
 
 
-        <activity android:name=".ui.chatroom.ChatRoomActivity">
+        <activity android:name="com.example.messenger.ui.chat.chatroom.ChatRoomActivity">
 <!--            <intent-filter>-->
 <!--                <action android:name="android.intent.action.MAIN" />-->
 <!--                <category android:name="android.intent.category.LAUNCHER" />-->
