+++ app/src/main/java/com/example/messenger/ui/chat/chatroom/adapter/MessageReceivedItemViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.messenger.ui.chatroom.adapter
+package com.example.messenger.ui.chat.chatroom.adapter
 
 import com.example.messenger.repository.model.chat.Message
