+++ app/src/main/java/com/example/messenger/ui/chat/chatroom/adapter/MessageReceivedViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.messenger.ui.chatroom.adapter
+package com.example.messenger.ui.chat.chatroom.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.example.messenger.base.BaseMessageViewHolder
