+++ app/src/main/java/com/example/messenger/ui/chat/chatroomlist/adapter/ChatRoomItemViewModel.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.example.messenger.ui.chatroomlist.adapter
+package com.example.messenger.ui.chat.chatroomlist.adapter
 
 import com.example.messenger.repository.model.chat.ChatRoom
