+++ app/src/main/res/layout/fragment_chat_room_list.xml
@@ -7,7 +7,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.example.messenger.ui.chatroomlist.ChatRoomListViewModel" />
+            type="com.example.messenger.ui.chat.chatroomlist.ChatRoomListViewModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
